Split class Chunk into multiple files
